+++ src/main/java/com/record/EmployeRecord.java
@@ -5,7 +5,7 @@ import java.util.Set;
 public record EmployeRecord(String              id, 
                             String              name, 
                             String              surname, 
-                            Double              salary, 
+                            Integer             salary, 
                             Set<String>         emails, 
                             Set<AddressRecord>  address) {
